# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: High-Risk Pregnancy Tracker API – Jazan Health Cluster 2026
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pregnancies:
    get:
      operationId: listPregnancies
      tags: [pregnancies]
      summary: List pregnancy cases with filters
      parameters:
        - name: patientId
          in: query
          required: false
          schema:
            type: integer
        - name: riskLevel
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: hospitalId
          in: query
          required: false
          schema:
            type: integer
        - name: sectorId
          in: query
          required: false
          schema:
            type: integer
        - name: compliance
          in: query
          required: false
          schema:
            type: string
            enum: [compliant, non_compliant, pending]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated list of pregnancies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PregnancyList"
    post:
      operationId: createPregnancy
      tags: [pregnancies]
      summary: Create a new pregnancy case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PregnancyInput"
      responses:
        "201":
          description: Created pregnancy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pregnancy"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pregnancies/{id}:
    get:
      operationId: getPregnancy
      tags: [pregnancies]
      summary: Get pregnancy case by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pregnancy details with patient info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PregnancyDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePregnancy
      tags: [pregnancies]
      summary: Update pregnancy case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PregnancyUpdate"
      responses:
        "200":
          description: Updated pregnancy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pregnancy"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Appointments ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ─── Reference entities ────────────────────────────────────────────────
...
    Patient:
      type: object
      required: [id, nationalId, nameAr, phone, healthCenterId, sectorId, createdAt]
      properties:
        id:
          type: integer
        nationalId:
          type: string
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        phone:
          type: string
        address:
          type: ["string", "null"]
        healthCenterId:
          type: integer
        healthCenterNameAr:
          type: ["string", "null"]
        sectorId:
          type: integer
        sectorNameAr:
          type: ["string", "null"]
        totalPregnancies:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Pregnancy:
      type: object
      required: [id, patientId, visitDate, gestationalAge, riskLevel, referralRecommendation, compliance, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientNameAr:
          type: ["string", "null"]
        patientNationalId:
          type: ["string", "null"]
        visitDate:
          type: string
        lmpDate:
          type: ["string", "null"]
        gestationalAge:
          type: ["integer", "null"]
          description: Gestational age in weeks
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        riskFactors:
          type: array
          items:
            type: string
        isVteHighRisk:
          type: boolean
        enoxaparinPrescribed:
          type: boolean
        doctorName:
          type: ["string", "null"]
        referralRecommendation:
          type: string
          enum: [follow_at_center, follow_at_hospital, transfer_kfch]
        referredHospitalId:
          type: ["integer", "null"]
        referredHospitalNameAr:
          type: ["string", "null"]
        appointmentDate:
          type: ["string", "null"]
        compliance:
          type: string
          enum: [compliant, non_compliant, pending]
          description: "compliant: <=2 working days, non_compliant: >2, pending: no appointment yet"
        workingDaysToAppointment:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        coordinatorClassification:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    PregnancyInput:
      type: object
      required: [patientId, visitDate, riskLevel, referralRecommendation]
      properties:
        patientId:
          type: integer
        visitDate:
          type: string
        lmpDate:
          type: ["string", "null"]
        gestationalAge:
          type: ["integer", "null"]
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        riskFactors:
          type: array
          items:
            type: string
        isVteHighRisk:
          type: boolean
        enoxaparinPrescribed:
          type: boolean
        doctorName:
          type: ["string", "null"]
        referralRecommendation:
          type: string
          enum: [follow_at_center, follow_at_hospital, transfer_kfch]
        referredHospitalId:
          type: ["integer", "null"]
        appointmentDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        coordinatorClassification:
          type: ["string", "null"]
    PregnancyUpdate:
      type: object
      properties:
        visitDate:
          type: string
        lmpDate:
          type: ["string", "null"]
        gestationalAge:
          type: ["integer", "null"]
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        riskFactors:
          type: array
          items:
            type: string
        isVteHighRisk:
          type: boolean
        enoxaparinPrescribed:
          type: boolean
        doctorName:
          type: ["string", "null"]
        referralRecommendation:
          type: string
          enum: [follow_at_center, follow_at_hospital, transfer_kfch]
        referredHospitalId:
          type: ["integer", "null"]
        appointmentDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        coordinatorClassification:
          type: ["string", "null"]
    PregnancyDetail:
      type: object
      required: [pregnancy, patient, appointments]
      properties:
        pregnancy:
          $ref: "#/components/schemas/Pregnancy"
        patient:
          $ref: "#/components/schemas/Patient"
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
    PregnancyList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Pregnancy"
        total:
          type: integer

    # ─── Appointment ──────────────────────────────────────────────────────
    Appointment:
      type: object
      required: [id, pregnancyId, hospitalId, appointmentDate, attended, createdAt]
      properties:
        id:
          type: integer
        pregnancyId:
          type: integer
        hospitalId:
          type: integer
        hospitalNameAr:
          type: ["string", "null"]
        appointmentDate:
          type: string
        attended:
          type: ["boolean", "null"]
          description: null = not yet, true = attended, false = missed
        attendanceNote:
          type: ["string", "null"]
        createdAt:
          type: string
...