# lib/api-zod/src/generated/types/pregnancyList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * High-Risk Pregnancy Tracker API – Jazan Health Cluster 2026
 * OpenAPI spec version: 0.1.0
 */
import type { Pregnancy } from './pregnancy';

export interface PregnancyList {
  items: Pregnancy[];
  total: number;
}